requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
deprecation==2.1.0
filelock==3.20.3
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
narwhals==2.15.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.0
postgrest==2.27.2
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
pypdf==6.6.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.27.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.1
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
storage3==2.27.2
streamlit==1.53.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
